Introduction to move
We put the words in a static variable, so that the closure will never
outlive them.

We need the move keyword to tell the closure to take ownership of the
reference.

We are not waiting for all threads to finish. Notice that the output
sometimes doesn't contain all the words. Proof that rust saved us from
doing something bad, since the closures indeed outlive the run function.