Introduction to iterators
Iterators are lazy, proof that the words are printed in correct order

What happens is that the map() part does not actually call the closure.

The closure is actually called later in the for loop. And since we wait
each handle before waiting the next, each thread is spawned after the
previous has been waited upon.