Let's take a step back
Ok, adding things to a vector was more complicated. Lets just try to
print the vector in each thread for start.

It is still not possible. Rust complains that the borrow doesn't live
long enough. And this is right! As we said the threads can be executed
whenever, even after the function is done. So they do need to be
borrowed for static.

In a sense this means that the vector has to live for as long as the
threads are alive. This reminds us of ownership. A value lives, as long
as the owner lives. Thus, in this case we have a kind of shared
ownership, where all the threads jointly own the vector.

Rust has a type for that!