Rust saves our ass again!
What went wrong? We used a shared ownership type, and rust still
complains!

Well, we didn't use a thread safe type! Reference counting obviously
needs to maintain some counter, and this means that loads and stores can
happen out of order and mess up the counter. If only we had something
that could load and store atomically...

Well, we do!